Extract timestamp parsing logic from process-event.ts
