+++ plugin-server/src/worker/ingestion/event-pipeline/2-prepareEventStep.ts
@@ -1,18 +1,16 @@
 import { PluginEvent } from '@posthog/plugin-scaffold'
-import { DateTime } from 'luxon'
 
+import { parseEventTimestamp } from '../timestamps'
 import { EventPipelineRunner, StepResult } from './runner'
 
 export async function prepareEventStep(runner: EventPipelineRunner, event: PluginEvent): Promise<StepResult> {
-    const { ip, site_url, team_id, now, sent_at, uuid } = event
-    const distinctId = String(event.distinct_id)
+    const { ip, site_url, team_id, uuid } = event
     const preIngestionEvent = await runner.hub.eventsProcessor.processEvent(
-        distinctId,
+        String(event.distinct_id),
         ip,
         event,
         team_id,
-        DateTime.fromISO(now),
-        sent_at ? DateTime.fromISO(sent_at) : null,
+        parseEventTimestamp(event, runner.hub.statsd),
         uuid! // it will throw if it's undefined,
     )
